requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
Deprecated==1.2.13
Django==3.2.9
django-environ==0.8.1
django-filter==21.1
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
jwcrypto==1.0
Markdown==3.3.6
oauthlib==3.1.1
pycparser==2.21
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.13.3
zipp==3.6.0
